LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/xslt.https.sub-expected.txt


PASS Same-Origin xslt 
FAIL undefined: sec-fetch-dest assert_equals: expected "xslt" but got ""
FAIL undefined: sec-fetch-mode assert_equals: expected "same-origin" but got ""
FAIL undefined: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS undefined: sec-fetch-user 
